Backend/admin/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
mysqlclient==2.1.0
django-mysql==3.9
django-cors-headers==3.8.0
pika==1.2.0